Delete disallows incorrect user from deleting from list

+++ api/controllers/booklistCont.js
@@ -38,26 +38,29 @@ export const addToList = (req, res)=>{
 }
 
 export const deleteFromList =(req, res)=>{
+
+    //CHECKS IF USER HAS AN ACCESS TOKEN IN THEIR COOKIES
     const token = req.cookies.access_token;
     if(!token) return res.status(401).json("User not authenticated");
-    
     jwt.verify(token, "jwtkey", (err, userInfo)=>{
         if(err) return res.status(403).json("Invalid web token");
-
-        const listId = req.params.id;
-        const q = `DELETE FROM books_in_booklists WHERE bookID = ? AND booklistID = ?`
+        // console.log(req.params.id);
+        // console.log(req.params.bookid);
+        // console.log(userInfo.id);
+        
+        //DELETES BOOK WHERE BOOKID AND BOOKLISTID MATCH, AS WELL AS MAKING SURE THE LOGGED USER IS THE OWNER OF THE LIST
+        const q = `DELETE FROM books_in_booklists WHERE bookID = ? AND booklistID = ? AND (SELECT userID FROM booklists WHERE id = ?) = ?`;
+        db.query(q, [req.params.bookid, req.params.id, req.params.id, userInfo.id], (err, data) => {
+            if(err) return res.status(403).json("You can't change another user's list");
+            return res.status(200).json("Book deleted from list");
+        })
     })
     
-    console.log(req.body);
-    console.log(req.params.id);
-    console.log(req.params.bookid);
+    
 
-    const q = `DELETE FROM books_in_booklists WHERE bookID = ? AND booklistID = ?`
+    
 
-    db.query(q, [req.params.bookid, req.params.id], (err, data) => {
-        if(err) return res.send(err);
-        return res.status(200).json("Book deleted");
-    })
+    
 }
